Move psammead-visually-hidden-text component out of legacy directory
