+++ src/app/components/Byline/index.tsx
@@ -5,7 +5,7 @@ import { jsx } from '@emotion/react';
 import React, { useContext, PropsWithChildren } from 'react';
 import pathOr from 'ramda/src/pathOr';
 import { ServiceContext } from '../../contexts/ServiceContext';
-import VisuallyHiddenText from '../../legacy/psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '../psammead-visually-hidden-text/src';
 import BylineCss from './index.styles';
 import { RightChevron, LeftChevron } from '../icons';
 import Text from '../Text';
