+++ src/app/components/Curation/CurationPromo/index.tsx
@@ -2,7 +2,7 @@
 import React, { useContext } from 'react';
 import moment from 'moment';
 import path from 'ramda/src/path';
-import VisuallyHiddenText from '#psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '#app/components/psammead-visually-hidden-text/src';
 import formatDuration from '#app/lib/utilities/formatDuration';
 import Promo from '#components/Promo';
 import { ServiceContext } from '../../../contexts/ServiceContext';
