+++ src/app/components/Curation/HierarchicalGrid/index.tsx
@@ -4,7 +4,7 @@ import { useContext } from 'react';
 import { css, jsx, Theme } from '@emotion/react';
 import moment from 'moment';
 import path from 'ramda/src/path';
-import VisuallyHiddenText from '../../../legacy/psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '../../psammead-visually-hidden-text/src';
 import formatDuration from '../../../lib/utilities/formatDuration';
 import Promo from '../../../legacy/components/Promo';
 import { DESKTOP, TABLET, MOBILE, SMALL } from './dataStructures';
