+++ src/app/legacy/components/SkipLinkWrapper/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { node, string, shape } from 'prop-types';
 import styled from '@emotion/styled';
 import detokenise from '#psammead/psammead-detokeniser/src';
-import { visuallyHiddenTextStyle } from '#psammead/psammead-visually-hidden-text/src';
+import { visuallyHiddenTextStyle } from '#app/components/psammead-visually-hidden-text/src';
 import { getSansBold } from '#psammead/psammead-styles/src/font-styles';
 import { GEL_BREVIER } from '#psammead/gel-foundations/src/typography';
