+++ src/app/legacy/containers/Copyright/index.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { string } from 'prop-types';
-import VisuallyHiddenText from '#psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '#app/components/psammead-visually-hidden-text/src';
 import Copyright from '#psammead/psammead-copyright/src';
 import { ServiceContext } from '../../../contexts/ServiceContext';
