+++ src/app/legacy/containers/EpisodeList/RecentAudioEpisodes/index.jsx
@@ -3,7 +3,7 @@ import React, { useContext } from 'react';
 import { arrayOf, shape, string, number } from 'prop-types';
 import styled from '@emotion/styled';
 import pathOr from 'ramda/src/pathOr';
-import VisuallyHiddenText from '#psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '#app/components/psammead-visually-hidden-text/src';
 import {
   formatDuration,
   formatUnixTimestamp,
