+++ src/app/legacy/containers/EpisodeList/RecentVideoEpisodes/index.jsx
@@ -4,7 +4,7 @@ import { useTheme } from '@emotion/react';
 import { arrayOf, shape, string, number } from 'prop-types';
 import styled from '@emotion/styled';
 import pathOr from 'ramda/src/pathOr';
-import VisuallyHiddenText from '#psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '#app/components/psammead-visually-hidden-text/src';
 import Timestamp from '#psammead/psammead-timestamp-container/src';
 import {
   formatDuration,
