+++ src/app/legacy/containers/VisuallyHiddenHeadline/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import path from 'ramda/src/path';
-import VisuallyHiddenText from '#psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '#app/components/psammead-visually-hidden-text/src';
 import { textDefaultPropTypes } from '#models/propTypes';
 import { headlineModelPropTypes } from '#models/propTypes/headline';
