+++ src/app/legacy/psammead/psammead-live-label/src/index.jsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import { node, bool, string, oneOf } from 'prop-types';
 import { getSansBold } from '#psammead/psammead-styles/src/font-styles';
 import { GEL_SPACING } from '#psammead/gel-foundations/src/spacings';
-import VisuallyHiddenText from '#psammead/psammead-visually-hidden-text/src';
+import VisuallyHiddenText from '#app/components/psammead-visually-hidden-text/src';
 
 const StyledSpan = styled.span`
   ${({ service }) => getSansBold(service)}
